git-svn-id: http://svn.mcdevzone.com/removr@54 3a220af4-2228-46fc-ad5e-55fe7f27f50b

+++ LevelEditor/LevelMap.m
@@ -15,8 +15,8 @@
 
 @synthesize dirty;
 
-- (UInt32)MapWidth  { return MAPWIDTH; }
-- (UInt32)MapHeight { return MAPHEIGHT; }
+- (NSInteger)MapWidth  { return MAPWIDTH; }
+- (NSInteger)MapHeight { return MAPHEIGHT; }
 
 // Access
